+++ astropy/modeling/tests/test_core.py
@@ -3,15 +3,14 @@
 from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
 
-import collections
-import inspect
-
 import pytest
 import numpy as np
 from numpy.testing.utils import assert_allclose
 from ..core import Model, InputParameterError, custom_model, render_model
 from ..parameters import Parameter
 from .. import models
+
+from ...utils.compat.funcsigs import signature
 
 
 class NonFittableModel(Model):
@@ -114,10 +113,10 @@ def test_custom_model_signature():
 
     assert model_a.param_names == ()
     assert model_a.n_inputs == 1
-    argspec = inspect.getargspec(model_a.__init__)
-    assert argspec.args == ['self']
-    argspec = inspect.getargspec(model_a.__call__)
-    assert argspec.args == ['self', 'x', 'model_set_axis']
+    sig = signature(model_a.__init__)
+    assert list(sig.parameters.keys()) == ['self', 'args', 'kwargs']
+    sig = signature(model_a.__call__)
+    assert list(sig.parameters.keys()) == ['self', 'x', 'model_set_axis']
 
     @custom_model
     def model_b(x, a=1, b=2):
@@ -125,11 +124,11 @@ def test_custom_model_signature():
 
     assert model_b.param_names == ('a', 'b')
     assert model_b.n_inputs == 1
-    argspec = inspect.getargspec(model_b.__init__)
-    assert argspec.args == ['self', 'a', 'b']
-    assert argspec.defaults == (1, 2)
-    argspec = inspect.getargspec(model_b.__call__)
-    assert argspec.args == ['self', 'x', 'model_set_axis']
+    sig = signature(model_b.__init__)
+    assert list(sig.parameters.keys()) == ['self', 'a', 'b', 'kwargs']
+    assert [x.default for x in sig.parameters.values()] == [sig.empty, 1, 2, sig.empty]
+    sig = signature(model_b.__call__)
+    assert list(sig.parameters.keys()) == ['self', 'x', 'model_set_axis']
 
     @custom_model
     def model_c(x, y, a=1, b=2):
@@ -137,11 +136,11 @@ def test_custom_model_signature():
 
     assert model_c.param_names == ('a', 'b')
     assert model_c.n_inputs == 2
-    argspec = inspect.getargspec(model_c.__init__)
-    assert argspec.args == ['self', 'a', 'b']
-    assert argspec.defaults == (1, 2)
-    argspec = inspect.getargspec(model_c.__call__)
-    assert argspec.args == ['self', 'x', 'y', 'model_set_axis']
+    sig = signature(model_c.__init__)
+    assert list(sig.parameters.keys()) == ['self', 'a', 'b', 'kwargs']
+    assert [x.default for x in sig.parameters.values()] == [sig.empty, 1, 2, sig.empty]
+    sig = signature(model_c.__call__)
+    assert list(sig.parameters.keys()) == ['self', 'x', 'y', 'model_set_axis']
 
 
 def test_custom_model_subclass():
@@ -162,10 +161,10 @@ def test_custom_model_subclass():
     assert b.a == 1
     assert b(1) == -1
 
-    argspec = inspect.getargspec(model_b.__init__)
-    assert argspec.args == ['self', 'a']
-    argspec = inspect.getargspec(model_b.__call__)
-    assert argspec.args == ['self', 'x', 'model_set_axis']
+    sig = signature(model_b.__init__)
+    assert list(sig.parameters.keys()) == ['self', 'a', 'kwargs']
+    sig = signature(model_b.__call__)
+    assert list(sig.parameters.keys()) == ['self', 'x', 'model_set_axis']
 
 
 def test_custom_model_parametrized_decorator():